Builds a schema using a simple methodology
Following this rule leads to effective naming of the schema, with role names that are contextualised within their relationships.

If a name cannot be found to concisely describe any of the 3 parts, then this indicates that the schema may need to be changed.

The rule:
A <relationship> has a <role> in the form of a <subject> (where <subject> is a concept, which could be an entity, relationship, or attribute)