Changes schema to add rules
Rules are useful to find the stations passed through when travelling from a to b along a route.
The idea is that once compute path has given a set of routes as a path, we can query for the connection from the start to the end of each leg, and examine the explanation to find those stops, in order.